SALTO-1342 - Move listUnresolvedReferences to workspace (#2110)
