+++ packages/workspace/index.ts
@@ -16,7 +16,7 @@
 import * as errors from './src/errors'
 import * as nacl from './src/workspace/nacl_files'
 import { Workspace, SourceFragment, StateRecency, loadWorkspace,
-  EnvironmentsSources, initWorkspace, WorkspaceComponents,
+  EnvironmentsSources, initWorkspace, WorkspaceComponents, UnresolvedElemIDs,
   COMMON_ENV_PREFIX } from './src/workspace/workspace'
 import * as hiddenValues from './src/workspace/hidden_values'
 import * as configSource from './src/workspace/config_source'
@@ -69,4 +69,5 @@ export {
   createElementSelectors,
   selectElementIdsByTraversal,
   ElementIDToValue,
+  UnresolvedElemIDs,
 }
